# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: KAYGO API - Plateforme de mise en relation logistique France-Martinique
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /auth/register:
    post:
      operationId: register
      tags: [auth]
      summary: Register a new user
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RegisterRequest"
      responses:
        "201":
          description: User created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthResponse"
        "400":
          description: Bad request
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /auth/login:
    post:
      operationId: login
      tags: [auth]
      summary: Login
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LoginRequest"
      responses:
        "200":
          description: Logged in
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthResponse"
        "401":
          description: Invalid credentials
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

  # ─── USERS ──────────────────────────────────────────────────────────────────
...
  /trips:
    get:
      operationId: listTrips
      tags: [trips]
      summary: List trips
      parameters:
        - name: status
          in: query
          schema:
            type: string
        - name: departureCity
          in: query
          schema:
            type: string
        - name: arrivalCity
          in: query
          schema:
            type: string
      responses:
        "200":
          description: List of trips
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TripsListResponse"
    post:
      operationId: createTrip
      tags: [trips]
      summary: Create a new trip
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateTripRequest"
      responses:
        "201":
          description: Trip created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TripResponse"
  /trips/{id}:
    get:
      operationId: getTrip
      tags: [trips]
      summary: Get a trip
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Trip
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TripResponse"
    patch:
      operationId: updateTrip
      tags: [trips]
      summary: Update trip status
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateTripRequest"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TripResponse"
  /trips/traveler/{travelerId}:
    get:
      operationId: getTravelerTrips
      tags: [trips]
      summary: Get trips for a traveler
      parameters:
        - name: travelerId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Trips list
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TripsListResponse"

  # ─── SHIPMENTS ──────────────────────────────────────────────────────────────
  /shipments:
    get:
      operationId: listShipments
      tags: [shipments]
      summary: List shipments
      parameters:
        - name: status
          in: query
          schema:
            type: string
        - name: senderId
          in: query
          schema:
            type: integer
      responses:
        "200":
          description: Shipments
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ShipmentsListResponse"
    post:
      operationId: createShipment
      tags: [shipments]
      summary: Create a new shipment request
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateShipmentRequest"
      responses:
        "201":
          description: Shipment created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ShipmentResponse"
  /shipments/{id}:
    get:
      operationId: getShipment
      tags: [shipments]
      summary: Get a shipment
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Shipment
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ShipmentResponse"
    patch:
      operationId: updateShipment
      tags: [shipments]
      summary: Update a shipment
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateShipmentRequest"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ShipmentResponse"
  /shipments/{id}/sender/{senderId}:
    get:
      operationId: getSenderShipments
      tags: [shipments]
      summary: Get shipments for a sender
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
        - name: senderId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Shipments
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ShipmentsListResponse"

  # ─── MATCHES ─────────────────────────────────────────────────────────────────
  /matches:
    get:
      operationId: listMatches
      tags: [matches]
      summary: List matches
      parameters:
        - name: adminStatus
          in: query
          schema:
            type: string
      responses:
        "200":
          description: Matches
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MatchesListResponse"
    post:
      operationId: createMatch
      tags: [matches]
      summary: Create a match (admin)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateMatchRequest"
      responses:
        "201":
          description: Match created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MatchResponse"
  /matches/{id}:
    get:
      operationId: getMatch
      tags: [matches]
      summary: Get a match
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Match
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MatchResponse"
    patch:
      operationId: updateMatch
      tags: [matches]
      summary: Update match status
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateMatchRequest"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MatchResponse"

  # ─── PAYMENTS ────────────────────────────────────────────────────────────────
...
  /proofs:
    post:
      operationId: createProof
      tags: [proofs]
      summary: Submit a delivery proof
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateProofRequest"
      responses:
        "201":
          description: Proof submitted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ProofResponse"
  /proofs/shipment/{shipmentId}:
    get:
      operationId: getShipmentProofs
      tags: [proofs]
      summary: Get proofs for a shipment
      parameters:
        - name: shipmentId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Proofs
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ProofsListResponse"

  # ─── NOTIFICATIONS ────────────────────────────────────────────────────────────
...
components:
  schemas:
...
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
        message:
          type: string
      required: [error]

    # ── Auth ──────────────────────────────────────────────────────────────────
    RegisterRequest:
      type: object
      properties:
        firstName:
          type: string
        lastName:
          type: string
        email:
          type: string
        phone:
          type: string
        password:
          type: string
        role:
          type: string
          enum: [traveler, sender, receiver, admin]
      required: [firstName, lastName, email, password, role]
    LoginRequest:
      type: object
      properties:
        email:
          type: string
        password:
          type: string
      required: [email, password]
    AuthResponse:
      type: object
      properties:
        token:
          type: string
        user:
          $ref: "#/components/schemas/User"
      required: [token, user]

    # ── User ──────────────────────────────────────────────────────────────────
    User:
      type: object
      properties:
        id:
          type: integer
        role:
          type: string
        firstName:
          type: string
        lastName:
          type: string
        email:
          type: string
        phone:
          type: string
          nullable: true
        avatarUrl:
          type: string
          nullable: true
        verificationStatus:
          type: string
        createdAt:
          type: string
      required: [id, role, firstName, lastName, email, verificationStatus, createdAt]
...
    Trip:
      type: object
      properties:
        id:
          type: integer
        travelerId:
          type: integer
        departureCity:
          type: string
        arrivalCity:
          type: string
        departureDate:
          type: string
        arrivalDate:
          type: string
        flightNumber:
          type: string
          nullable: true
        transportType:
          type: string
        baggageTotalKg:
          type: number
        baggageUsedKg:
          type: number
        baggageFreeKg:
          type: number
        baggageType:
          type: string
        acceptsExtraBaggage:
          type: boolean
        acceptedCategories:
          type: array
          items:
            type: string
        minReward:
          type: number
        notes:
          type: string
          nullable: true
        status:
          type: string
        createdAt:
          type: string
        travelerName:
          type: string
          nullable: true
      required: [id, travelerId, departureCity, arrivalCity, departureDate, arrivalDate, baggageTotalKg, baggageUsedKg, baggageFreeKg, minReward, status, createdAt]
    TripResponse:
      type: object
      properties:
        trip:
          $ref: "#/components/schemas/Trip"
      required: [trip]
    TripsListResponse:
      type: object
      properties:
        trips:
          type: array
          items:
            $ref: "#/components/schemas/Trip"
        total:
          type: integer
      required: [trips, total]
    CreateTripRequest:
      type: object
      properties:
        travelerId:
          type: integer
        departureCity:
          type: string
        arrivalCity:
          type: string
        departureDate:
          type: string
        arrivalDate:
          type: string
        flightNumber:
          type: string
        transportType:
          type: string
        baggageTotalKg:
          type: number
        baggageUsedKg:
          type: number
        baggageFreeKg:
          type: number
        baggageType:
          type: string
        acceptsExtraBaggage:
          type: boolean
        acceptedCategories:
          type: array
          items:
            type: string
        minReward:
          type: number
        notes:
          type: string
      required: [travelerId, departureCity, arrivalCity, departureDate, arrivalDate, baggageTotalKg, baggageUsedKg, baggageFreeKg, minReward]
    UpdateTripRequest:
      type: object
      properties:
        status:
          type: string
        baggageUsedKg:
          type: number
        baggageFreeKg:
          type: number
        notes:
          type: string

    # ── Shipment ──────────────────────────────────────────────────────────────
    Shipment:
      type: object
      properties:
        id:
          type: integer
        senderId:
          type: integer
        receiverName:
          type: string
        receiverPhone:
          type: string
        departureCity:
          type: string
        arrivalCity:
          type: string
        weightKg:
          type: number
        dimensions:
          type: string
          nullable: true
        category:
          type: string
        description:
          type: string
          nullable: true
        declaredValue:
          type: number
          nullable: true
        photoUrls:
          type: array
          items:
            type: string
        urgencyLevel:
          type: string
        pickupOption:
          type: boolean
        deliveryOption:
          type: boolean
        pickupAddress:
          type: string
          nullable: true
        deliveryAddress:
          type: string
          nullable: true
        status:
          type: string
        createdAt:
          type: string
        senderName:
          type: string
          nullable: true
      required: [id, senderId, receiverName, receiverPhone, departureCity, arrivalCity, weightKg, category, status, createdAt]
    ShipmentResponse:
      type: object
      properties:
        shipment:
          $ref: "#/components/schemas/Shipment"
      required: [shipment]
    ShipmentsListResponse:
      type: object
      properties:
        shipments:
          type: array
          items:
            $ref: "#/components/schemas/Shipment"
        total:
          type: integer
      required: [shipments, total]
    CreateShipmentRequest:
      type: object
      properties:
        senderId:
          type: integer
        receiverName:
          type: string
        receiverPhone:
          type: string
        departureCity:
          type: string
        arrivalCity:
          type: string
        weightKg:
          type: number
        dimensions:
          type: string
        category:
          type: string
        description:
          type: string
        declaredValue:
          type: number
        photoUrls:
          type: array
          items:
            type: string
        urgencyLevel:
          type: string
        pickupOption:
          type: boolean
        deliveryOption:
          type: boolean
        pickupAddress:
          type: string
        deliveryAddress:
          type: string
      required: [senderId, receiverName, receiverPhone, departureCity, arrivalCity, weightKg, category]
    UpdateShipmentRequest:
      type: object
      properties:
        status:
          type: string
        photoUrls:
          type: array
          items:
            type: string
        notes:
          type: string

    # ── Match ─────────────────────────────────────────────────────────────────
    Match:
      type: object
      properties:
        id:
          type: integer
        shipmentId:
          type: integer
        tripId:
          type: integer
        proposedReward:
          type: number
        platformFee:
          type: number
        serviceFee:
          type: number
        pickupFee:
          type: number
        deliveryFee:
          type: number
        totalPrice:
          type: number
        adminStatus:
          type: string
        travelerStatus:
          type: string
        senderStatus:
          type: string
        createdAt:
          type: string
        shipment:
          $ref: "#/components/schemas/Shipment"
          nullable: true
        trip:
          $ref: "#/components/schemas/Trip"
          nullable: true
      required: [id, shipmentId, tripId, proposedReward, platformFee, serviceFee, pickupFee, deliveryFee, totalPrice, adminStatus, travelerStatus, senderStatus, createdAt]
    MatchResponse:
      type: object
      properties:
        match:
          $ref: "#/components/schemas/Match"
      required: [match]
    MatchesListResponse:
      type: object
      properties:
        matches:
          type: array
          items:
            $ref: "#/components/schemas/Match"
        total:
          type: integer
      required: [matches, total]
    CreateMatchRequest:
      type: object
      properties:
        shipmentId:
          type: integer
        tripId:
          type: integer
        proposedReward:
          type: number
        platformFee:
          type: number
        serviceFee:
          type: number
        pickupFee:
          type: number
        deliveryFee:
          type: number
        totalPrice:
          type: number
      required: [shipmentId, tripId, proposedReward, platformFee, serviceFee, pickupFee, deliveryFee, totalPrice]
    UpdateMatchRequest:
      type: object
      properties:
        adminStatus:
          type: string
        travelerStatus:
          type: string
        senderStatus:
          type: string

    # ── Payment ───────────────────────────────────────────────────────────────
...
    Proof:
      type: object
      properties:
        id:
          type: integer
        shipmentId:
          type: integer
        proofType:
          type: string
        photoUrl:
          type: string
          nullable: true
        note:
          type: string
          nullable: true
        deliveryCode:
          type: string
          nullable: true
        createdAt:
          type: string
      required: [id, shipmentId, proofType, createdAt]
    ProofResponse:
      type: object
      properties:
        proof:
          $ref: "#/components/schemas/Proof"
      required: [proof]
    ProofsListResponse:
      type: object
      properties:
        proofs:
          type: array
          items:
            $ref: "#/components/schemas/Proof"
      required: [proofs]
    CreateProofRequest:
      type: object
      properties:
        shipmentId:
          type: integer
        proofType:
          type: string
        photoUrl:
          type: string
        note:
          type: string
        deliveryCode:
          type: string
      required: [shipmentId, proofType]

    # ── Notification ──────────────────────────────────────────────────────────
...