# lib/api-zod/src/generated/types/matchesListResponse.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * KAYGO API - Plateforme de mise en relation logistique France-Martinique
 * OpenAPI spec version: 0.1.0
 */
import type { Match } from "./match";

export interface MatchesListResponse {
  matches: Match[];
  total: number;
}