# lib/api-zod/src/generated/types/updateTripRequest.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * KAYGO API - Plateforme de mise en relation logistique France-Martinique
 * OpenAPI spec version: 0.1.0
 */

export interface UpdateTripRequest {
  status?: string;
  baggageUsedKg?: number;
  baggageFreeKg?: number;
  notes?: string;
}